Added "Ignite floor"
Added an effect that sets fire to the ground below you. If you're in a car it will ignite the floor in a 5x5 area while as if you are on foot it will only start one fire directly below you. I'm still unsure about the size of both fires because it's not really a problem unless you're standing on top of something that can spread the fire...